+++ app/src/main/java/com/example/android/sunshine/app/MainActivity.java
@@ -1,24 +1,11 @@
 package com.example.android.sunshine.app;
 
 import android.support.v7.app.ActionBarActivity;
-import android.support.v4.app.Fragment;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.util.ArrayList;
+import com.example.android.sunshine.app.Fragments.ForcastFragment;
 
 
 public class MainActivity extends ActionBarActivity {
