+++ app/src/main/res/layout/fragment_detail.xml
@@ -9,7 +9,8 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="sample"
+        android:layout_marginLeft="15dp"
         android:textAlignment="center"
-        android:layout_marginTop="220dp"/>
+        android:layout_marginTop="20dp"/>
 
 </LinearLayout>
